+++ app/models/carto/visualization.rb
@@ -57,7 +57,6 @@ class Carto::Visualization < ActiveRecord::Base
   belongs_to :user, inverse_of: :visualizations, select: Carto::User::DEFAULT_SELECT
   belongs_to :full_user, class_name: Carto::User, foreign_key: :user_id, primary_key: :id, inverse_of: :visualizations, readonly: true
 
-  belongs_to :user_table, class_name: Carto::UserTable, primary_key: :map_id, foreign_key: :map_id, inverse_of: :visualization
   has_one :vis_category, class_name: Carto::Category, primary_key: :category, foreign_key: :id
 
   belongs_to :permission, inverse_of: :visualization, dependent: :destroy
@@ -162,12 +161,21 @@ class Carto::Visualization < ActiveRecord::Base
     map.layers.select { |l| l.data_readable_by?(user) }
   end
 
+  def layers_with_data_and_common_shared_data_readable_by(user)
+    return [] unless map
+    map.layers.select { |l| l.data_and_common_shared_data_readable_by?(user) }
+  end
+
   def related_tables
     @related_tables ||= get_related_tables
   end
 
   def related_tables_readable_by(user)
     layers_with_data_readable_by(user).map { |l| l.user_tables_readable_by(user) }.flatten.uniq
+  end
+
+  def related_tables_and_common_shared_data_readable_by(user)
+    layers_with_data_and_common_shared_data_readable_by(user).map { |l| l.user_tables_and_common_shared_data_readable_by(user) }.flatten.uniq
   end
 
   def related_canonical_visualizations