+++ app/models/map.rb
@@ -141,7 +141,12 @@ class Map < Sequel::Model
   end
 
   def can_add_layer(user)
-    return false if self.user.max_layers && self.user.max_layers <= data_layers.count
+    empty_dataset_name = Cartodb.config[:shared_empty_dataset_name]
+    layers_on_map = data_layers
+    layer_count = layers_on_map.count
+    layer_count -= 1 if layers_on_map.any?{ |layer| layer.options["table_name"] == empty_dataset_name }
+
+    return false if self.user.max_layers && self.user.max_layers <= layer_count
     return false if self.user.viewer
 
     current_vis = visualizations.first