+++ services/importer/spec/acceptance/geojson_spec.rb
@@ -21,6 +21,12 @@ describe 'geojson regression tests' do
   include AcceptanceHelpers
   include_context "no stats"
 
+  let :ogr2ogr2_options do
+    {
+      ogr2ogr_binary: Cartodb.config[:ogr2ogr]['binary']
+    }
+  end
+
   before(:all) do
     @user = create_user
     @user.save
@@ -39,6 +45,7 @@ describe 'geojson regression tests' do
                                log: CartoDB::Importer2::Doubles::Log.new(@user),
                                user: @user
                              })
+    runner.loader_options = ogr2ogr2_options
     runner.run
   end
 
@@ -52,6 +59,7 @@ describe 'geojson regression tests' do
                                log: CartoDB::Importer2::Doubles::Log.new(@user),
                                user: @user
                              })
+    runner.loader_options = ogr2ogr2_options
     runner.run
   end
 
@@ -64,6 +72,7 @@ describe 'geojson regression tests' do
                                log: CartoDB::Importer2::Doubles::Log.new(@user),
                                user: @user
                              })
+    runner.loader_options = ogr2ogr2_options
     runner.run
 
     runner.results.first.error_code.should eq 1002