+++ spec/models/map_spec.rb
@@ -553,6 +553,24 @@ describe Map do
 
         @map.can_add_layer(@user).should eq false
       end
+
+      it 'should take shared_empty_dataset into account' do
+        empty_dataset_name = Cartodb.config[:shared_empty_dataset_name]
+        @map = Map.create(user_id: @user.id, table_id: @table.id)
+        @map.reload
+
+        @layer = Layer.create(kind: 'carto', options: { :table_name => empty_dataset_name })
+        @map.add_layer(@layer)
+        @map.reload
+
+        @user.max_layers.times do
+          @map.can_add_layer(@user).should eq true
+          @map.add_layer(Layer.new(kind: 'carto'))
+          @map.reload
+        end
+
+        @map.can_add_layer(@user).should eq false
+      end
     end
   end
 end